Added files via upload
bar1.html is a simple working example of a bar graph populated into a div element.
complex_bar_graph.html is a non-working file containing bar graph code and some of the notes to get it working for our needs.
Source is https://bost.ocks.org/mike/bar. It is a 3-page tutorial for that leads up to the code from complex_bar_graph. I will have more time to work on this Thursday. Or we could try something with bar1.html. The hard part is vertical bars, axes, bells and whistles.