Added files via upload
File holds a javascript function that takes a custom array and a CSS class name. Class name is connected to a div. Function is called at the bottom of the script element, and it prints the graph of values in the custom array. Source is https://bost.ocks.org/mike/bar tutorial.